requirements.txt
requests
pillow
discord-webhook
